@@ -12,32 +12,34 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.0.0",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
     "@angular/elements": "^7.0.0",
-    "@angular/forms": "~7.0.0",
+    "@angular/forms": "^8.0.0",
     "@angular/http": "~7.0.0",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
+    "@angular/material": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/platform-server": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@webcomponents/custom-elements": "^1.2.0",
     "@webcomponents/webcomponentsjs": "^2.1.2",
     "concat": "^1.0.3",
     "core-js": "^2.5.4",
     "ngx-pagination": "^5.0.0",
     "rxjs": "~6.3.3",
+    "subsink": "^1.0.1",
     "zone.js": "~0.9.0"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^7.0.0",
-    "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.7",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
+    "@angular-builders/custom-webpack": "^8.0.0",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
@@ -52,6 +54,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "3.4.5"
   }
 }
